requirements.txt
Flask==3.0.0
Flask-Mail==0.9.1
requests==2.31.0
python-dotenv==1.0.1
gunicorn==21.2.0